Big refactor - created new files + moved code around
